A couple of fixes for text, validation and a new user role

+++ app/Http/Requests/UpdateDisc.php
@@ -23,7 +23,7 @@ class UpdateDisc extends FormRequest
     {
         return [
             'condition' => 'string|nullable',
-            'offer_price' => 'integer',
+            'offer_price' => 'numeric',
         ];
     }
 }
